add conditions on net.py

+++ src/ip.py
@@ -1,17 +1,18 @@
 #!/opt/ipmanager/venv/bin/python3
-from getpass import getpass
 from mysql.connector import connect, Error
+import ipaddress
+
+
+#This is function for check a conditions
 
 def valid_ip(ip):
-    parts = ip.split(".")
-    if len(parts) != 4:
+    try:
+        ipaddress.ip_address(ip)
+        return True
+    except:
         return False
-    for item in parts:
-        if not 0 <= int(item) <= 255:
-            return False
-    return True
 
-def show_ip(ip):
+def check_ip(ip):
     valid_ip(ip)
     if valid_ip(ip) is True:
         query = f"select * FROM ipaddresses WHERE ip = '{ip}' "
@@ -19,13 +20,33 @@ def show_ip(ip):
         cursor = conn.cursor()
         cursor.execute(query)
         result = cursor.fetchall()
+        if len(result) == 0:
+            return True
+        else:
+            return False
+    else:
+        return False
+    
+    
+#This is function for interaction with database
+
+def show_ip(ip):
+    valid_ip(ip)
+    check_ip(ip)
+    if valid_ip(ip) is True and check_ip(ip) is False:
+        query = f"select * FROM ipaddresses WHERE ip = '{ip}' "
+        conn = connect(host="127.0.0.1", user="admin", password="secret", database="ipmanager_db")
+        cursor = conn.cursor()
+        cursor.execute(query)
+        result = cursor.fetchall()
         return result
     else:
-        return {"message": "ip-addres is invalid"}
+        return {"message": "ip-addres is invalid or doesnt exist"}
 
 def add_ip(ip, used, comment):
     valid_ip(ip)
-    if valid_ip(ip) is True:
+    check_ip(ip)
+    if valid_ip(ip) is True and check_ip(ip) is True:
         query = f"insert into ipaddresses (`ip`, `used`, `comment`) value ('{ip}', '{used}', '{comment}')"
         conn = connect(host="127.0.0.1", user="admin", password="secret", database="ipmanager_db")
         cursor = conn.cursor()
@@ -34,11 +55,12 @@ def add_ip(ip, used, comment):
         result = cursor.fetchall()
         return result
     else:
-        return {"message": "ip-addres is invalid"}
+        return {"message": "ip-addres is invalid or already exist"}
 
 def del_ip(ip):
     valid_ip(ip)
-    if valid_ip(ip) is True:
+    check_ip(ip)
+    if valid_ip(ip) is True and check_ip(ip) is False:
         query = f"delete from ipaddresses where ip = '{ip}'"
         conn = connect(host="127.0.0.1", user="admin", password="secret", database="ipmanager_db")
         cursor = conn.cursor()
@@ -47,11 +69,12 @@ def del_ip(ip):
         result = cursor.fetchall()
         return result
     else:
-        return {"message": "ip-addres is invalid"}
+        return {"message": "ip-addres is invalid or already deleted"}
 
 def upd_ip(ip, used, comment):
     valid_ip(ip)
-    if validIP(ip) is True:
+    check_ip(ip)
+    if valid_ip(ip) is True and check_ip(ip) is False:
         query = f"update ipaddresses set `used` = '{used}', `comment` = '{comment}' where ip = '{ip}'"
         conn = connect(host="127.0.0.1", user="admin", password="secret", database="ipmanager_db")
         cursor = conn.cursor()
@@ -60,7 +83,7 @@ def upd_ip(ip, used, comment):
         result = cursor.fetchall()
         return result
     else:
-        return {"message": "ip-addres is invalid"}
+        return {"message": "ip-addres is invalid or doesnt exist"}
 
 
 def search_ip():
